+++ src/main/kotlin/glsl/plugin/psi/named/types/builtins/GlslVector.kt
@@ -3,6 +3,7 @@ package glsl.plugin.psi.named.types.builtins
 import com.intellij.lang.ASTNode
 import com.intellij.psi.tree.IElementType
 import com.intellij.psi.util.elementType
+import glsl.GlslTypes.*
 import glsl.data.GlslDefinitions
 import glsl.data.GlslErrorMessages
 import glsl.plugin.psi.named.GlslNamedElement
@@ -10,6 +11,7 @@ import glsl.plugin.psi.named.GlslNamedType
 import glsl.plugin.psi.named.GlslNamedTypeImpl
 import glsl.plugin.psi.named.GlslNamedVariable
 import glsl.plugin.utils.GlslBuiltinUtils
+import glsl.plugin.utils.GlslUtils.createScalarTypeElement
 import glsl.psi.interfaces.GlslBuiltinTypeVector
 
 
@@ -22,7 +24,7 @@ abstract class GlslVector(node: ASTNode) : GlslNamedTypeImpl(node), GlslBuiltinT
      *
      */
     override fun getPsi(): GlslBuiltinTypeVector {
-        return this as GlslBuiltinTypeVector
+        return firstChild as GlslBuiltinTypeVector
     }
 
     /**
@@ -38,6 +40,21 @@ abstract class GlslVector(node: ASTNode) : GlslNamedTypeImpl(node), GlslBuiltinT
      */
     override fun getStructMember(memberName: String): GlslNamedVariable? {
         return getStructMembers().find { it.name == memberName }
+    }
+
+    /**
+     *
+     */
+    override fun getScalarType(): GlslNamedType? {
+        val typeText = name?.first()
+        return when (typeText) {
+            'v', 'f' -> createScalarTypeElement(FLOAT, "float")
+            'i' -> createScalarTypeElement(INT, "int")
+            'u' -> createScalarTypeElement(UINT, "uint")
+            'b' -> createScalarTypeElement(BOOL, "bool")
+            'd' -> createScalarTypeElement(DOUBLE, "double")
+            else -> null
+        }
     }
 
     /**
@@ -76,20 +93,5 @@ abstract class GlslVector(node: ASTNode) : GlslNamedTypeImpl(node), GlslBuiltinT
         if (other == null) return false
         val implicitConversions = GlslDefinitions.VECTORS[other]
         return implicitConversions?.contains(elementType) ?: false
-    }
-
-    /**
-     *
-     */
-    private fun getVectorComponentType(): String {
-        val typeText = name?.first()
-        return when (typeText) {
-            'v', 'f' -> "float"
-            'd' -> "double"
-            'u' -> "uint"
-            'i' -> "int"
-            'b' -> "bool"
-            else -> ""
-        }
     }
 }