+++ src/test/kotlin/GlslDummyTest.kt
@@ -8,12 +8,11 @@ class GlslDummyTest : BasePlatformTestCase() {
     }
 
     fun testDummy() {
-        myFixture.configureByFiles("dummy.glsl")
-//        myFixture.checkHighlighting(false, false, false)
-        myFixture.enableInspections(GlslInspectionIncompatibleType())
-        myFixture.testHighlighting(false, false, false)
-//        val reference = myFixture.getReferenceAtCaretPosition("dummy.glsl")
-//        val resolve = reference?.resolve()
-//        assertNotNull(resolve)
+//        myFixture.configureByFiles("dummy.glsl")
+//        myFixture.enableInspections(GlslInspectionIncompatibleType())
+//        myFixture.testHighlighting(false, false, false)
+        val reference = myFixture.getReferenceAtCaretPosition("dummy.glsl")
+        val resolve = reference?.resolve()
+        assertNotNull(resolve)
     }
 }